Created signup
Activation Keys are generated by MD5 hashing the Unique ID

Password, since admin access is required to add users, are generated by
shuffling the key and MD5 hashing the output.

SMTP settings need to be set in the main controller and an email must be 
set in the config.

Both emails are sent independently.  

Signed-off-by: lgdolman <l.g.dolman@gmail.com>